Move Ollama code editor to intermediate layer
